some more examples

+++ pixi/src/main/scala/pixigraphics/WithTint.scala
@@ -6,6 +6,6 @@ import scala.scalajs.js
 @js.native
 trait WithTint extends js.Object {
 
-  var tint: Int = js.native
+  var tint: Double = js.native
 
 }
